requirements.txt
django
Jinja2
requests
parsel
django-redis
Pillow
PyMySQL
python-alipay-sdk